third_party/blink/web_tests/fast/hidpi/image-srcset-density-and-width-descriptors-expected.txt

CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/green-400-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/green-400-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/green-400-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/green-400-px-square.png"
PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("foo").clientWidth==100 is true
This test passes if the image below is a blue square, since the srcset resource doesn't load, as it has both 'w' and 'x' descriptors.
